requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
python-decouple==3.8
django-rest-knox==4.2.0
pyotp==2.9.0
sendgrid==6.11.0
firebase-admin==6.2.0  
python-jose[cryptography]==3.3.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
requests==2.31.0
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0